Refactor tests packages
