+++ tests/profiles_enabled/telemetry/test_telemetry.py
