+++ tests/profiles_enabled/test_rpc_client.py
